+++ vllm/distributed/__init__.py
@@ -1,2 +1,3 @@
 from vllm.distributed.parallel_state import *
 from vllm.distributed.communication_op import *
+from vllm.distributed.utils import *
