+++ vllm/model_executor/layers/linear.py
@@ -10,7 +10,7 @@ from vllm.distributed import (
     tensor_model_parallel_all_gather, tensor_model_parallel_all_reduce)
 from vllm.distributed import (
     get_tensor_model_parallel_rank, get_tensor_model_parallel_world_size)
-from vllm.utils import (
+from vllm.distributed import (
     divide, split_tensor_along_last_dim)
 from vllm.model_executor.utils import set_weight_attrs
